memory efficient metrics

+++ bench/metric_bench.cpp
@@ -5,7 +5,7 @@
 static void bench_original_metrics(benchmark::State& state) 
 {
     using namespace haisu::metric;
-    monitor<int, 16> mon;
+    monitor<int8_t, 16> mon;
     while (state.KeepRunning())
     {
         mon.start(1);
@@ -30,7 +30,7 @@ static void bench_original_metrics(benchmark::State& state)
 static void bench_trie_metrics(benchmark::State& state) 
 {
     namespace met = haisu::metric2;
-    met::monitor<int> mon;
+    met::monitor<int8_t> mon;
     while (state.KeepRunning())
     {
         mon.start(1);
@@ -55,7 +55,7 @@ static void bench_trie_metrics(benchmark::State& state)
 static void bench_original_metrics_shallow(benchmark::State& state) 
 {
     using namespace haisu::metric;
-    monitor<int, 16> mon;
+    monitor<int8_t, 16> mon;
     while (state.KeepRunning())
     {
         mon.start(1);
@@ -66,7 +66,7 @@ static void bench_original_metrics_shallow(benchmark::State& state)
 static void bench_trie_metrics_shallow(benchmark::State& state) 
 {
     namespace met = haisu::metric2;
-    met::monitor<int> mon;
+    met::monitor<int8_t> mon;
     while (state.KeepRunning())
     {
         mon.start(1);
@@ -77,7 +77,7 @@ static void bench_trie_metrics_shallow(benchmark::State& state)
 static void bench_original_metrics_flat_hierarchy(benchmark::State& state) 
 {
     using namespace haisu::metric;
-    monitor<int, 16> mon;
+    monitor<int8_t, 16> mon;
     mon.start(1);
     while (state.KeepRunning())
     {
@@ -95,7 +95,7 @@ static void bench_original_metrics_flat_hierarchy(benchmark::State& state)
 static void bench_trie_metrics_flat_hierarchy(benchmark::State& state) 
 {
     namespace met = haisu::metric2;
-    met::monitor<int> mon;
+    met::monitor<int8_t> mon;
     mon.start(1);
     while (state.KeepRunning())
     {
